+++ src/components/iplDialog.vue
@@ -5,8 +5,14 @@
         @cancel.prevent="onCancel"
         @click.self="onClick"
     >
+        <div class="header">
+            <slot name="header" />
+        </div>
         <div class="content">
             <slot />
+        </div>
+        <div class="footer">
+            <slot name="footer" />
         </div>
     </dialog>
 </template>
@@ -88,6 +94,18 @@ dialog {
     top: 50%;
     transform: translate(-50%, -50%);
     border-radius: constants.$border-radius-outer;
+    max-height: calc(100dvh - 16px);
+}
+
+// Touchscreen devices tend to have random UI elements loitering around the bottom of the screen
+// (e.g. the little bar you swipe up from to return to the home screen; I don't know what it's called.
+// I can see it on my iPad and my Pixel, so it isn't unique to one operating system.)
+// This media query addresses that by adding more empty space below dialogs on touchscreens.
+@media (pointer: coarse) {
+    dialog {
+        max-height: calc(100dvh - 32px);
+        top: calc(50% - 8px);
+    }
 }
 
 @include dialogs.dialog-common-animation;